+++ web/app/page.tsx
@@ -1,12 +1,13 @@
 "use client"
 
 import { useState } from 'react';
-import { format } from 'date-fns';
 
 import CenterContainer from './components/CenterContainer';
 import DateForm from './components/DateForm';
-import DiaryForm, { DiaryObjType } from './components/DiaryForm';
+import DiaryForm from './components/DiaryForm';
 import SuccessMessageToast from './components/SuccessMessageToast';
+
+import { DiaryObjType, useDiary } from './useDiary';
 
 export default function Home() {
   const [date, setDate] = useState<Date | null>(null) // 日付フィールドの値
@@ -24,6 +25,7 @@ export default function Home() {
   async function handleDateSubmit() {
     if (!date) return
   
+    const { getDiaryByDate } = useDiary()
     const diary = await getDiaryByDate(date)
     setDiaryObj({
       text: diary.is_text ? diary.text : "",
@@ -39,6 +41,7 @@ export default function Home() {
   async function handleDiaryUpdate() {
     if (!diaryObj.date) return
 
+    const { updateDiary } = useDiary()
     const message = await updateDiary(diaryObj.text, diaryObj.date)
 
     setMessageOnUpdated(message)
@@ -66,42 +69,3 @@ export default function Home() {
   )
 }
 
-async function getDiaryByDate(date: Date): Promise<{ text: string, is_text: boolean }> {
-  const formatedDateStr = format(date, "yyyyMMdd")
-
-  const res = await fetch('https://fetch-diary-uezai2z2jq-uc.a.run.app', {
-    method: "POST",
-    headers: {
-			"Content-Type": "application/json"
-		},
-		body: JSON.stringify({ "date": formatedDateStr })
-  })
-
-  if (!res.ok) {
-    // This will activate the closest `error.js` Error Boundary
-    throw new Error('Failed to fetch diary data')
-  }
-
-  const diaryData = await res.json()
-  return diaryData
-}
-
-async function updateDiary(diary: string, date: Date) {
-  const formatedDateStr = format(date, "yyyyMMdd")
-
-  const res = await fetch('https://update-diary-uezai2z2jq-uc.a.run.app', {
-    method: "POST",
-    headers: {
-			"Content-Type": "application/json"
-		},
-		body: JSON.stringify({ "date": formatedDateStr, "text": diary })
-  })
-
-  if (!res.ok) {
-    // This will activate the closest `error.js` Error Boundary
-    throw new Error('Failed to update diary data')
-  }
-
-  const message = await res.text()
-  return message
-}
